move GeoIP functions into it's own class
extract the geoip database .gz in-game
