+++ OpenRA.Mods.Common/Widgets/Logic/ServerBrowserLogic.cs
@@ -252,7 +252,7 @@ namespace OpenRA.Mods.Common.Widgets.Logic
 					var location = item.GetOrNull<LabelWidget>("LOCATION");
 					if (location != null)
 					{
-						var cachedServerLocation = LobbyUtils.LookupCountry(game.Address.Split(':')[0]);
+						var cachedServerLocation = GeoIP.LookupCountry(game.Address.Split(':')[0]);
 						location.GetText = () => cachedServerLocation;
 						location.GetColor = () => !compatible ? Color.DarkGray : !canJoin ? Color.LightGray : location.TextColor;
 					}
